add base classes, impl data, domain
