+++ CleanArchExample/app/src/main/kotlin/com/example/cleanarch/di/Injectable.kt
@@ -1,0 +1,62 @@
+package com.example.cleanarch.di
+
+import android.app.Activity
+import android.app.Application
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentActivity
+import androidx.fragment.app.FragmentManager
+import dagger.android.AndroidInjection
+import dagger.android.support.AndroidSupportInjection
+import dagger.android.support.HasSupportFragmentInjector
+
+interface Injectable
+
+fun Application.applyAutoInjector() = registerActivityLifecycleCallbacks(
+    object : Application.ActivityLifecycleCallbacks {
+
+        override fun onActivityCreated(activity: Activity, savedInstanceState: Bundle?) {
+            handleActivity(activity)
+        }
+
+        override fun onActivityStarted(activity: Activity) {
+
+        }
+
+        override fun onActivityResumed(activity: Activity) {
+
+        }
+
+        override fun onActivityPaused(activity: Activity) {
+
+        }
+
+        override fun onActivityStopped(activity: Activity) {
+
+        }
+
+        override fun onActivitySaveInstanceState(activity: Activity, outState: Bundle?) {
+
+        }
+
+        override fun onActivityDestroyed(activity: Activity) {
+
+        }
+    })
+
+private fun handleActivity(activity: Activity) {
+    if (activity is Injectable || activity is HasSupportFragmentInjector) {
+        AndroidInjection.inject(activity)
+    }
+    if (activity is FragmentActivity) {
+        activity.supportFragmentManager.registerFragmentLifecycleCallbacks(
+            object : FragmentManager.FragmentLifecycleCallbacks() {
+                override fun onFragmentCreated(fm: FragmentManager, f: Fragment, s: Bundle?) {
+                    if (f is Injectable) {
+                        AndroidSupportInjection.inject(f)
+                    }
+                }
+            }, true
+        )
+    }
+}
