--- CleanArchExample/data/src/main/kotlin/com/example/cleanarch/di/ApiModules.kt
@@ -1,8 +1,0 @@
-package com.example.cleanarch.di
-
-import dagger.Module
-
-/* module for API connection(e.g retrofit2, http3, etc..) */
-@Module
-class ApiModules {
-}
