+++ CleanArchExample/data/src/main/kotlin/com/example/data/utils/di/ApiModules.kt
@@ -1,0 +1,8 @@
+package com.example.data.utils.di
+
+import dagger.Module
+
+/* module for API connection(e.g retrofit2, http3, etc..) */
+@Module
+class ApiModules {
+}
