--- CleanArchExample/domain/src/main/kotlin/com/example/cleanarch/utils/commons/ExecutionThreads.kt
@@ -1,8 +1,0 @@
-package com.example.cleanarch.utils.commons
-
-import io.reactivex.Scheduler
-
-interface ExecutionThreads {
-    fun io(): Scheduler
-    fun ui(): Scheduler
-}
