--- CleanArchExample/domain/src/main/kotlin/com/example/cleanarch/utils/commons/InputUseCase.kt
@@ -1,4 +1,0 @@
-package com.example.cleanarch.utils.commons
-
-abstract class InputUseCase<in Q : UseCase.RequestValue, T : Throwable>(executionThreads: ExecutionThreads) :
-    IoUseCase<Q, UseCase.NoResponseValue, T>(executionThreads)
