--- CleanArchExample/domain/src/main/kotlin/com/example/cleanarch/utils/commons/IoUseCase.kt
@@ -1,4 +1,0 @@
-package com.example.cleanarch.utils.commons
-
-abstract class IoUseCase<in Q : UseCase.RequestValue, R : UseCase.ResponseValue, T : Throwable>(executionThreads: ExecutionThreads) :
-    UseCase<Q, R, T>(executionThreads)
