--- CleanArchExample/domain/src/main/kotlin/com/example/cleanarch/utils/commons/NoIoUseCase.kt
@@ -1,9 +1,0 @@
-package com.example.cleanarch.utils.commons
-
-import io.reactivex.Observable
-
-abstract class NoIoUseCase<T : Throwable>(executionThreads: ExecutionThreads) :
-    IoUseCase<UseCase.NoRequestValue, UseCase.NoResponseValue, T>(executionThreads) {
-    protected abstract fun execute(): Observable<NoResponseValue>
-    override fun execute(requestValue: NoRequestValue): Observable<NoResponseValue> = execute()
-}
