--- CleanArchExample/domain/src/main/kotlin/com/example/cleanarch/utils/di/UseCaseModules.kt
@@ -1,9 +1,0 @@
-package com.example.cleanarch.utils.di
-
-import dagger.Module
-
-@Module
-abstract class UseCaseModules {
-//    @Binds
-//    abstract fun bindGetIndustries(getIndustriesImpl: GetIndustriesImpl): GetIndustries
-}
