+++ CleanArchExample/domain/src/main/kotlin/com/example/domain/dayofweek/entity/DayOfWeek.kt
@@ -1,0 +1,39 @@
+package com.example.domain.dayofweek.entity
+
+enum class DayOfWeek {
+    /**
+     * The singleton instance for the day-of-week of Monday.
+     * This has the numeric value of `1`.
+     */
+    MONDAY,
+    /**
+     * The singleton instance for the day-of-week of Tuesday.
+     * This has the numeric value of `2`.
+     */
+    TUESDAY,
+    /**
+     * The singleton instance for the day-of-week of Wednesday.
+     * This has the numeric value of `3`.
+     */
+    WEDNESDAY,
+    /**
+     * The singleton instance for the day-of-week of Thursday.
+     * This has the numeric value of `4`.
+     */
+    THURSDAY,
+    /**
+     * The singleton instance for the day-of-week of Friday.
+     * This has the numeric value of `5`.
+     */
+    FRIDAY,
+    /**
+     * The singleton instance for the day-of-week of Saturday.
+     * This has the numeric value of `6`.
+     */
+    SATURDAY,
+    /**
+     * The singleton instance for the day-of-week of Sunday.
+     * This has the numeric value of `7`.
+     */
+    SUNDAY;
+}
