+++ CleanArchExample/domain/src/main/kotlin/com/example/domain/dayofweek/entity/Language.kt
@@ -1,0 +1,7 @@
+package com.example.domain.dayofweek.entity
+
+enum class Language {
+    English,
+    German,
+    Japanese;
+}
