+++ CleanArchExample/domain/src/main/kotlin/com/example/domain/repository/DayOfWeekRepository.kt
@@ -1,0 +1,9 @@
+package com.example.domain.repository
+
+import com.example.domain.dayofweek.entity.DayOfWeek
+import com.example.domain.dayofweek.entity.Language
+
+interface DayOfWeekRepository {
+    fun getTodaysDayOfWeek(): DayOfWeek
+    fun getDayOfWeekWords(language: Language): Map<DayOfWeek, String>
+}
