+++ CleanArchExample/domain/src/main/kotlin/com/example/domain/utils/commons/ExecutionThreads.kt
@@ -1,0 +1,8 @@
+package com.example.domain.utils.commons
+
+import io.reactivex.Scheduler
+
+interface ExecutionThreads {
+    fun io(): Scheduler
+    fun ui(): Scheduler
+}
