+++ CleanArchExample/domain/src/main/kotlin/com/example/domain/utils/commons/InputUseCase.kt
@@ -1,0 +1,4 @@
+package com.example.domain.utils.commons
+
+abstract class InputUseCase<in Q : UseCase.RequestValue, T : Throwable>(executionThreads: ExecutionThreads) :
+    IoUseCase<Q, UseCase.NoResponseValue, T>(executionThreads)
