+++ CleanArchExample/domain/src/main/kotlin/com/example/domain/utils/commons/IoUseCase.kt
@@ -1,0 +1,4 @@
+package com.example.domain.utils.commons
+
+abstract class IoUseCase<in Q : UseCase.RequestValue, R : UseCase.ResponseValue, T : Throwable>(executionThreads: ExecutionThreads) :
+    UseCase<Q, R, T>(executionThreads)
