+++ CleanArchExample/domain/src/main/kotlin/com/example/domain/utils/di/UseCaseModules.kt
@@ -1,0 +1,9 @@
+package com.example.domain.utils.di
+
+import dagger.Module
+
+@Module
+abstract class UseCaseModules {
+//    @Binds
+//    abstract fun bindGetIndustries(getIndustriesImpl: GetIndustriesImpl): GetIndustries
+}
