+++ CleanArchExample/presentation/src/main/AndroidManifest.xml
@@ -1,2 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.example.presentation"/>
+          package="com.example.presentation">
+
+    <application>
+        <activity android:name=".presenter.main.CleanArchMainActivity">
+        </activity>
+    </application>
+
+</manifest>
