--- CleanArchExample/presentation/src/main/kotlin/com/example/cleanarch/utils/annotations/ActivityScope.kt
@@ -1,3 +1,0 @@
-package com.example.cleanarch.utils.annotations
-
-annotation class ActivityScope
