--- CleanArchExample/presentation/src/main/kotlin/com/example/cleanarch/utils/annotations/ViewModelKey.kt
@@ -1,12 +1,0 @@
-package com.example.cleanarch.utils.annotations
-
-import androidx.lifecycle.ViewModel
-import dagger.MapKey
-import kotlin.reflect.KClass
-
-
-@MustBeDocumented
-@Target(AnnotationTarget.FUNCTION)
-@Retention(AnnotationRetention.RUNTIME)
-@MapKey
-internal annotation class ViewModelKey(val value: KClass<out ViewModel>)
