--- CleanArchExample/presentation/src/main/kotlin/com/example/cleanarch/utils/commons/BasePresenter.kt
@@ -1,5 +1,0 @@
-package com.example.cleanarch.utils.commons
-
-interface BasePresenter {
-
-}
