--- CleanArchExample/presentation/src/main/kotlin/com/example/cleanarch/utils/commons/BaseView.kt
@@ -1,4 +1,0 @@
-package com.example.cleanarch.utils.commons
-
-interface BaseView {
-}
