--- CleanArchExample/presentation/src/main/kotlin/com/example/cleanarch/utils/extensions/Activity.kt
@@ -1,7 +1,0 @@
-package com.example.cleanarch.utils.extensions
-
-import android.app.Activity
-
-inline fun <reified T> Activity.lazyWithExtras(key: String): Lazy<T> {
-    return lazy { intent.extras.get(key) as T }
-}
