--- CleanArchExample/presentation/src/main/kotlin/com/example/cleanarch/utils/extensions/Fragment.kt
@@ -1,7 +1,0 @@
-package com.example.cleanarch.utils.extensions
-
-import androidx.fragment.app.Fragment
-
-inline fun <reified T> Fragment.lazyWithArgs(key: String): Lazy<T> {
-    return lazy { arguments!!.get(key) as T }
-}
