--- CleanArchExample/presentation/src/main/kotlin/com/example/cleanarch/utils/extensions/LiveDataReactiveStreams.kt
@@ -1,7 +1,0 @@
-package com.example.cleanarch.utils.extensions
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.LiveDataReactiveStreams
-import org.reactivestreams.Publisher
-
-fun <T> Publisher<T>.toLiveData() = LiveDataReactiveStreams.fromPublisher(this) as LiveData<T>
