+++ CleanArchExample/presentation/src/main/kotlin/com/example/presentation/presenter/main/CleanArchMainContract.kt
@@ -1,0 +1,15 @@
+package com.example.presentation.presenter.main
+
+import com.example.presentation.utils.commons.BaseView
+import com.example.presentation.utils.commons.BasePresenter
+
+interface CleanArchMainContract {
+
+    interface View : BaseView {
+
+    }
+
+    interface Presenter : BasePresenter {
+
+    }
+}
