+++ CleanArchExample/presentation/src/main/kotlin/com/example/presentation/utils/annotations/ActivityScope.kt
@@ -1,0 +1,3 @@
+package com.example.presentation.utils.annotations
+
+annotation class ActivityScope
