+++ CleanArchExample/presentation/src/main/kotlin/com/example/presentation/utils/annotations/ViewModelKey.kt
@@ -1,0 +1,12 @@
+package com.example.presentation.utils.annotations
+
+import androidx.lifecycle.ViewModel
+import dagger.MapKey
+import kotlin.reflect.KClass
+
+
+@MustBeDocumented
+@Target(AnnotationTarget.FUNCTION)
+@Retention(AnnotationRetention.RUNTIME)
+@MapKey
+internal annotation class ViewModelKey(val value: KClass<out ViewModel>)
