+++ CleanArchExample/presentation/src/main/kotlin/com/example/presentation/utils/commons/BasePresenter.kt
@@ -1,0 +1,5 @@
+package com.example.presentation.utils.commons
+
+interface BasePresenter {
+
+}
