+++ CleanArchExample/presentation/src/main/kotlin/com/example/presentation/utils/commons/BaseView.kt
@@ -1,0 +1,4 @@
+package com.example.presentation.utils.commons
+
+interface BaseView {
+}
