+++ CleanArchExample/presentation/src/main/kotlin/com/example/presentation/utils/extensions/Activity.kt
@@ -1,0 +1,7 @@
+package com.example.presentation.utils.extensions
+
+import android.app.Activity
+
+inline fun <reified T> Activity.lazyWithExtras(key: String): Lazy<T> {
+    return lazy { intent.extras.get(key) as T }
+}
