+++ CleanArchExample/presentation/src/main/kotlin/com/example/presentation/utils/extensions/Fragment.kt
@@ -1,0 +1,7 @@
+package com.example.presentation.utils.extensions
+
+import androidx.fragment.app.Fragment
+
+inline fun <reified T> Fragment.lazyWithArgs(key: String): Lazy<T> {
+    return lazy { arguments!!.get(key) as T }
+}
