+++ CleanArchExample/presentation/src/main/kotlin/com/example/presentation/utils/extensions/LiveDataReactiveStreams.kt
@@ -1,0 +1,7 @@
+package com.example.presentation.utils.extensions
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.LiveDataReactiveStreams
+import org.reactivestreams.Publisher
+
+fun <T> Publisher<T>.toLiveData() = LiveDataReactiveStreams.fromPublisher(this) as LiveData<T>
